### 2023-03-10 #5246 @editorialbot
[PRE REVIEW]: yam: Yet another monitoring tool using correlations of ambient noise
**Submitting author:** @trichter (<a href="http://orcid.org/0000-0002-8378-559X">Tom Eulenfeld</a>)
**Repository:** https://github.com/trichter/yam
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.1
