### 2023-03-13 #5246 @elbeejay
@trichter I have two requests for you at the moment:

1. Do you have any suggestions for reviewers of your work? JOSS maintains a [list of potential reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822), but we are also very receptive to any suggestions you may be have who have expertise in your field.
2. Can you add a section to the paper titled "Statement of need"? Hopefully this is self-explanatory, but we require that section so that authors can explicitly outline the need their software package is satisfying. This is designed to make clear to the reader if the paper they are reading is going to be useful for them.

I'll be taking a closer look at the paper and repository this week and will let you know if there are any other obvious changes I think need to happen.

Thanks,
Jay
