### 2023-03-13 #5246 @trichter
@elbeejay Thank you for editing the submission.

I added a section "Statement of need". The article is relatively short. It is basically just this one section. I hope this is OK.

In the list I found 14 reviewers with topic seismology. Of these I think jkmacc-LANL, d-chambers, seisman, core-man, calum-chamberlain could be potential reviewers for this submission.
