### 2023-03-13 #5246 @trichter
Not sure if something is wrong with the bibliography, because `Sens-Schönfelder, Christoph, & Eulenfeld, T. (2019)` is displayed differently than the other references (in the text with the first name, in the bibliography with the first name spelled out).
