### 2023-03-14 #5246 @elbeejay
> Not sure if something is wrong with the bibliography, because Sens-Schönfelder, Christoph, & Eulenfeld, T. (2019) is displayed differently than the other references (in the text with the first name, in the bibliography with the first name spelled out).

Thanks for pointing this out. I think it is because the `.bib` entries are not consistent, the 2006 reference abbreviates Christoph to C., while the 2014 reference uses Christoph, and the 2019 reference also uses Christoph. In the paper the 2006 reference is rendered with the C., the 2014 reference is truncated to "et al." and the 2019 reference has the full first name spelled out. I suggest you make these bibliography entries consistent and I would expect the inconsistency in the text to go away.
