### 2023-03-15 3717667
consistent bibliography entries
