+++ joss/paper.bib
@@ -1,5 +1,5 @@
 @article{Snieder2002,
-  author   = {Snieder, Roel and Gr{\^{e}}t, Alexandre and Douma, Huub and Scales, John},
+  author   = {Snieder, R. and Gr{\^{e}}t, A. and Douma, H. and Scales, J.},
   journal  = {Science},
   title    = {Coda wave interferometry for estimating nonlinear behavior in seismic velocity.},
   year     = {2002},
@@ -12,7 +12,7 @@
 }
 
 @article{Weaver2002,
-  author   = {Weaver, Richard and Lobkis, Oleg},
+  author   = {Weaver, R. and Lobkis, O.},
   journal  = {Ultrasonics},
   title    = {On the emergence of the {Green's} function in the correlations of a diffuse field: Pulse-echo using thermal phonons},
   year     = {2002},
@@ -48,7 +48,7 @@
 }
 
 @article{Richter2014,
-  author   = {Richter, Tom and Sens-Sch{\"{o}}nfelder, Christoph and Kind, Rainer and Asch, G{\"{u}}nter},
+  author   = {Richter, T. and Sens-Sch{\"{o}}nfelder, C. and Kind, R. and Asch, G.},
   journal  = {Journal of Geophysical Research: Solid Earth},
   title    = {Comprehensive observation and modeling of earthquake and temperature-related seismic velocity changes in northern {Chile} with passive image interferometry},
   year     = {2014},
@@ -60,7 +60,7 @@
 }
 
 @article{SensSchoenfelder2019,
-  author  = {Sens-Sch{\"{o}}nfelder, Christoph and Eulenfeld, Tom},
+  author  = {Sens-Sch{\"{o}}nfelder, C. and Eulenfeld, T.},
   journal = {Physical Review Letters},
   title   = {Probing the in situ elastic nonlinearity of rocks with earth tides and seismic noise},
   year    = {2019},
@@ -83,7 +83,7 @@
 }
 
 @Article{msnoise,
-  author       = {T. Lecocq and C. Caudron and F. Brenguier},
+  author       = {Lecocq, T. and Caudron, C. and Brenguier, F.},
   journal      = {Seismological Research Letters},
   title        = {{MSNoise}, a {Python} package for monitoring seismic velocity changes using ambient seismic noise},
   year         = {2014},
