### 2023-03-15 #5246 @hfmark
Hi @elbeejay - I'd be happy to review this. I haven't worked on `obspy` directly but, full disclosure, I have contributed to one of @trichter's other projects, `rf`. That was over a year ago, but if it raises any red flags for you then of course I won't take this on.
