### 2023-03-15 #5246 @brmather
Hi @elbeejay - I don't think I'm a good person to review this. I don't have any experience in seismic processing, so I would not know how the repo contributes and adds to existing tools. My apologies for this!
