### 2023-03-15 #5246 @elbeejay
@ThomasMGeo are you able and willing to review this submission to JOSS?
