### 2023-03-21 #5267 @ThomasMGeo
@trichter is Yam designed only as a command line tool? Going through the tutorial notebook, having every line need a `!` is somewhat of an odd design choice for the python scientific ecosystem.

This is not a dealbreaker for this submission, but I would like to see more information about this, and utility on the readme itself.
