### 2023-03-21 #5267 @trichter
@ThomasMGeo Yes, yam is primarily designed as an end-to-end command line tool. You are right, the first example is basically calling commands with `!` and displaying images. For me the Jupyter notebook was the best tool for demonstration purposes. The second, more advanced example showcases a bit more Python code.
The readme is very short and basic information can be found in the documentation ("how to use yam").
