### 2023-03-21 #5267 @ThomasMGeo
I would add some more information to the readme, and make it more explicit that it is a command line tool.
