### 2023-04-10 #5267 @ThomasMGeo
@trichter , looks great! Happy to sign off the documentation parts when you add a few things to the readme:

1. Does not mention that it's a command line tool on the readme, would put that in.
2. You mention 'yet another', I would probably link to 1-2 other projects with similar functionality. This would also cover the statement of need (why did you develop this?).
3. I would personally add the installation instructions to the readme as well (projects like pandas do this). This is not a requirement.
