### 2023-04-17 #5267 @hfmark
Ok, a few thoughts for @trichter:

1/ for installation instructions, could you clarify the list of optional dependencies? I'm thinking `tqdm` should be in there. And if you're suggesting including matplotlib, numpy, and scipy in the conda command maybe they should be listed too (though they're also dependencies for obspy so I suppose you could also leave them out of the conda command).

2/ For the tutorial section in the documentation, note first that the date for the PATCX data should be 2010-02-13 for what gets downloaded with `yam create --tutorial` rather than 2010-02-03.
Second, I think it's a little misleading that the tutorial code block is a list of commands that one *could* run, but you can't actually run them in sequence as listed on the page. For example, there's a plot command for stacks before the stack is made.
Third issue with those tutorial commands: after running the suggested installation instructions none of the plotting functions work for me. Seems to be a Qt problem which makes me wonder if there's a missing dependency or a setting that needs adjusting? Though plotting does work in the jupyter notebooks when I run those locally.
And lastly, I'd suggest adding a direct link to the notebook example near that documentation tutorial rather than just stating it's linked in the readme of the repository.

(re: the notebooks, I tried running the first one for the inter-station cross-correlations and after an hour+ of downloading data and 2 hours of working on correlations with -n 2 it overheated my laptop. It did run on my desktop, but takes a long time as far as code demos go, which is one reason why I think it would be helpful to have the simpler tutorial be more functional)

3/ community guidelines are mostly taken care of by pointing to the obspy forum, but it wouldn't hurt to add a line to the repo readme noting that people can contribute via issues/pull requests.

4/ the statement of need: to me this comes across much more clearly in the documentation than in the paper. Stating that MSNoise uses a sql database and yam is more concise without one doesn't strike me as great motivation; what the documentation suggested to me is that yam is more convenient for offline use and that the human-readable config files and obspy-readable output files were a key difference. Making the case for your code takes on a little extra importance, perhaps, when you put "yet another" in the name :)

5/ last but not least, to hop on earlier comments about the command line vs python question: I agree that you could state a bit more forcefully that yam is intended as a command line tool. The paper summary says it's a command line package, but the second paragraph of the statement of need says it's a python package. I get that it's sort of both (and the mix of traditional and semi-pythonic command line arguments illustrates that too) but from a user perspective maybe you need a different term? Say, "a python-based command line module"?
