### 2023-04-17 #5267 @hfmark
I also just noticed a tiny error in the second notebook example, opened an issue for it (trichter/notebookds#1)
