### 2023-04-17 #5267 @hfmark
and also! Just figured out the plotting issue I mentioned previously, I just didn't realize that the default is to save figures in `plots/`. For the purposes of a tutorial I'd suggest mentioning that `--show` brings up the interactive window, in case some of your future users are, like me, not super observant when it comes to these things.
