### 2023-04-20 25e89ac
doc: clarify deps
