### 2023-04-20 b62fc1c
docs: mention plots directory and show flag in tutorial
