### 2023-04-20 c0d0a51
doc: directly link tutorial notebooks
