+++ README.md
@@ -15,6 +15,7 @@ Yet another monitoring tool using correlations of ambient noise
 * Showcase / advanced tutorial: Monitoring seismic velocity -> [notebook](http://nbviewer.jupyter.org/github/trichter/notebooks/blob/master/yam_velocity_variations_patcx/processing_patcx.ipynb)
 * Yam poster at EGU 2018, doi: [10.6084/m9.figshare.16976017](https://doi.org/10.6084/m9.figshare.16976017)
 * Get help or discuss in the ObsPy forum, please use the [ObsPy Related Projects category](https://discourse.obspy.org/c/obspy-related-projects)
+* You are welcome to contribute code by opening pull requests or report bugs in the issue tracker
 
 
 #### Citation
