### 2023-04-20 2514ae4
readme: mention cli character
