### 2023-04-20 fd144ab
readme: mention install instructions
