+++ joss/paper.md
@@ -22,7 +22,8 @@ bibliography: paper.bib
 
 # Summary
 
-By calculating the cross-correlation of seismic noise between two stations, it is possible to retrieve the Green's function between the receivers. In addition, the cross-correlation function can be used to monitor changes in the seismic velocity in the subsurface. We present ``yam`` -- a command line package for calculating cross-correlations and relative velocity changes.
+By calculating the cross-correlation of seismic noise between two stations, it is possible to retrieve the Green's function between the receivers. In addition, the cross-correlation function can be used to monitor changes in the seismic velocity in the subsurface.
+We present ``yam`` -- a Python-based command line package for calculating cross-correlations and relative velocity changes.
 
 # Statement of need
 
@@ -30,19 +31,18 @@ Monitoring with cross-correlations of ambient noise is a popular technique for i
 The method combines two concepts -- Green's function retrieval between two receivers by cross-correlating an isotropic, homogeneous noise field recorded at the two receivers [@Weaver2002; @Shapiro2004] and velocity monitoring using coda wave interferometry [@Snieder2002].
 For monitoring, the condition of homogeneity and isotropy of the noise field can be relaxed in in favor of the more convenient condition of constancy of the noise sources.
 
-``yam`` is an ObsPy based [@obspy] Python package for correlating seismic recordings of ambient vibrations and for the monitoring of relative seismic velocity changes.
-The package is designed to be run from the command line for off-line use, but also includes capabilities for processing a continuously growing dataset.
-``yam`` does not rely on a database, but checks on the fly which results already exist and which results still need to be calculated.
+``yam`` is an ObsPy-based [@obspy] Python command line package for correlating seismic recordings of ambient vibrations and for the monitoring of relative seismic velocity changes.
+Another popular package for this task is MSNoise [@msnoise], which is especially useful for large datasets and continuous monitoring because it uses a sqlite or mysql database.
+``yam``, contrary to MSNoise, is designed for off-line use, but also includes capabilities for processing a continuously growing dataset.
+``yam`` does not rely on a database, but checks on the fly which results already exist and which results still have to be calculated.
 Cross-correlations are written to HDF5 files using the ObsPy plugin obspyh5.
 This makes it easy to access the correlation data after computation using ObsPy's ``read()`` function.
 Correlations can also be exported to various seismic formats to allow the determination of surface wave dispersion curves.
 The analysis of changes in the cross-correlation functions is implemented using the stretching procedure.
-One of the strengths of the code is the configuration, which is declared in a simple but heavily commented JSON file.
+One of the strengths of the code is the configuration, which is declared in a simple but heavily commented JSON file, unlike to a web interface used with MSNoise.
 It is possible to declare similar configurations without explicit repetition.
 A possible use case is to reprocess an entire dataset in a different frequency band or to stretch with the same parameters using a different time window.
 Parts of this code have been successfully used in @Richter2014 to estimate velocity changes induced by ground shaking and thermal stresses, and in @SensSchoenfelder2019 to estimate velocity changes induced by tidal stresses.
-
-Another popular package for velocity monitoring using correlations of ambient vibrations is MSNoise [@msnoise], which is especially useful for large datasets and continuous monitoring because it uses a sqlite or mysql database. The ``yam`` package does not use this concept in favor of a concise code base.
 
 ``yam`` is intended for both researchers and graduate students of seismology.
 The package can be installed from PyPI; online documentation and tutorials are available on the project website.
