### 2023-04-20 #5267 @trichter
@ThomasMGeo , thank you again for your review!

> @trichter , looks great! Happy to sign off the documentation parts when you add a few things to the readme:
>
> 1. Does not mention that it's a command line tool on the readme, would put that in.

Thanks! I added a sentence mentioning the CLI character of yam to the readme.

> 2. You mention 'yet another', I would probably link to 1-2 other projects with similar functionality. This would also cover the statement of need (why did you develop this?).

I reworked the whole paragraph in the paper to make the differences to MSNoise clearer, which is the the most popular package for this task.

> 3. I would personally add the installation instructions to the readme as well (projects like pandas do this). This is not a requirement.

I would prefer to maintain a single location for install instructions. I added text to the readme referring to the docs for install instructions.

@hfmark, thank you again for your review!

> 1/ for installation instructions, could you clarify the list of optional dependencies? I'm thinking `tqdm` should be in there. And if you're suggesting including matplotlib, numpy, and scipy in the conda command maybe they should be listed too (though they're also dependencies for obspy so I suppose you could also leave them out of the conda command).

`tqdm` is a hard dependency specified in setup.py and used in the commands.py module. I added it to the list. Only IPython and cartopy are optional and already listed as optional. As you suggested, I removed the mpl, numpy and scipy packages from the conda command as they are hard deps of obspy.

> 2/ For the tutorial section in the documentation, note first that the date for the PATCX data should be 2010-02-13 for what gets downloaded with `yam create --tutorial` rather than 2010-02-03.

I don't really understand what you mean. The data range is 2010-02-01 to 2010-02-14 00:00 with some artificial gaps. Could you explain your thoughts?

>Second, I think it's a little misleading that the tutorial code block is a list of commands that one _could_ run, but you can't actually run them in sequence as listed on the page. For example, there's a plot command for stacks before the stack is made.

For me, the commands are working in the given order. Could you try again? Note, that for correlation `c1` the stacking over 1 day is already performed with the `correlate` command. This is not the case for the `cauto` correlation configuration. This behavior is mentioned in the comments of the created template configuration `conf.json`.

>~Third issue with those tutorial commands: after running the suggested installation instructions none of the plotting functions work for me. Seems to be a Qt problem which makes me wonder if there's a missing dependency or a setting that needs adjusting? Though plotting does work in the jupyter notebooks when I run those locally.~ _nvm, this wasn't a real problem_ And lastly, I'd suggest adding a direct link to the notebook example near that documentation tutorial rather than just stating it's linked in the readme of the repository.

Done. Thanks for the suggestion.

> 3/ community guidelines are mostly taken care of by pointing to the obspy forum, but it wouldn't hurt to add a line to the repo readme noting that people can contribute via issues/pull requests.

Thanks. Done.

> 4/ the statement of need: to me this comes across much more clearly in the documentation than in the paper. Stating that MSNoise uses a sql database and yam is more concise without one doesn't strike me as great motivation; what the documentation suggested to me is that yam is more convenient for offline use and that the human-readable config files and obspy-readable output files were a key difference. Making the case for your code takes on a little extra importance, perhaps, when you put "yet another" in the name :)

I reworked the whole paragraph in the paper to make the differences to MSNoise clearer.

> 5/ last but not least, to hop on earlier comments about the command line vs python question: I agree that you could state a bit more forcefully that yam is intended as a command line tool. The paper summary says it's a command line package, but the second paragraph of the statement of need says it's a python package. I get that it's sort of both (and the mix of traditional and semi-pythonic command line arguments illustrates that too) but from a user perspective maybe you need a different term? Say, "a python-based command line module"?

Thanks for this suggestion. I updated the readme and the paper.

>and also! Just figured out the plotting issue I mentioned previously, I just didn't realize that the default is to save figures in plots/. For the purposes of a tutorial I'd suggest mentioning that --show brings up the interactive window, in case some of your future users are, like me, not super observant when it comes to these things.

That is definitely a good point and it is now stated.
