### 2023-04-20 #5267 @hfmark
@trichter I definitely messed something up when I first tried the documentation tutorial - probably tried to run the first notebook locally in the same place. Sorry! The `example_data/` therefore didn't initially include a file for PATCX on the listed date so I assumed it was a typo, and possibly the json file was wrong for the same reason, but starting from scratch in a clean directory it worked (except for one actual typo - `yam plot cauto_t2` should I think be `yam plot cauto_s2` in the 3rd-to-last line?).

The edits to the paper, readme, and docs all look good to me. I'll check off the rest of my list - as far as I'm concerned this is good to go!
