### 2023-04-21 2afbeab
docs: fix one cmd in tutorial
