### 2023-04-21 #5267 @trichter
> yam plot cauto_t2 should I think be yam plot cauto_s2 in the 3rd-to-last line?).

Good catch, I fixed it.
