### 2023-04-25 #5267 @hemmelig
**Review of `yam`: Yet Another Monitoring tool using correlations of ambient noise**

**General**
1. License file is present, but years are not up to date (not sure if important) and there is a mix of instances of the old license and the new one in the headers of files.
2. You might consider adding a `CITATION.cff` file (either to the Zenodo registration of the code, or the JOSS paper when finalised).
3. How does this benchmark against the alternatives out there? I see it is parallelised internally—how do you specify the number of jobs it can spawn? From the source code, it seems you can pass in this as an argument on the command line, but this isn't reference anywhere I can find.

**Manuscript**
I think it would be useful, for the layperson, to highlight what exactly is the value of monitoring relative velocity changes. I also don’t understand in what sense this competes with MSNoise? What do you mean by off-line?

You might specify a reference for the stretching procedure.

I like the human-readable, static config file approach.

And are graduate students not researchers? Perhaps you mean it is designed for research and teaching purposes.

**Code**
**Platform specs**
**OS:** macOS 12.5.1
**Arch:** arm64 (M1 chip)
**Python:** 3.9.12

**Download, installation, testing**
I was able to download and install `yam` from both the source repository and PyPI. Running the command-line test utility (`yam-runtests`) failed on 1 test. This seemed to be solely timing related, and not to do with the functionality of the code itself. By commenting out lines 173 and 174 and re-installing, I was able to run the tests with no further issues.

**Tutorials**
I’ll preface this by saying I am not good at following instructions in a linear fashion and will often glaze over some things, run commands, then go back and try figure out why things happened the way they did….

1.	Clarify at the top of the tutorial that the commands are being run in a terminal
2.	Perhaps be explicit and include two additional lines at the top of the box:
a.	`mkdir yam_tutorial; cd yam_tutorial`
b.	`yam create --tutorial`
I’ll admit I glazed over the ‘into an empty directory’ and just leapt straight in.
4.	`yam plot stations` does not work on Mac M1 when building from source. This is due to `cartopy` not being listed in the requirements. `cartopy` cannot be installed from PyPI (on my M1 system at least), but it does work with `conda`. There are no issues if you the install instructions as specified It might be worth adding a note explicitly stating `cartopy` must be installed via conda, even if building from source?
5.	Was very easy to grok how to set things up to run on a locally hosted data set, specifying an archive structure etc. I really like that you can create a collection of configuration recipes without needing to maintain multiple config JSONs. The keys can become a little unwieldy, but after a couple of runs editing things I’m pretty happy with them.
6.	I understood what was happening a lot better when I ran through each section in conjunction with the `conf.json` file. I would recommend people have it open while running through the example sequence of commands.
7.	I echo @hfmark’s comment—" the human-readable config files and obspy-readable output files [a]re a key difference” that should be highlighted strongly in the manuscript.

The test source code contained references to a different package, by the way:

```
yam/tests/__init__.py: Tests for the rf package.
```

I'm looking forward to finding time to apply `yam` to a more extensive swathe of the seismic data I've worked with!

I'll finalise the checklist, but I'm in principle ok to say this is ready to go, once you've had a chance to get back to me on some of the above points. Thanks for your patience!
