+++ yam/tests/test_correlate.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 Tom Eulenfeld, MIT license
+# Copyright 2017-2023 Tom Eulenfeld, MIT license
 import unittest
 import sys
 import warnings
