+++ yam/util.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 Tom Eulenfeld, MIT license
+# Copyright 2017-2023 Tom Eulenfeld, MIT license
 """Utility functions"""
 
 from importlib import import_module
