### 2023-04-26 f9939c3
relax timing test
