### 2023-04-26 6bbe010
mention njobs flag in docs
