### 2023-04-26 #5267 @trichter
> **General**
>
> 1. License file is present, but years are not up to date (not sure if important) and there is a mix of instances of the old license and the new one in the headers of files.
>

Thank you for the very helpful review! I fixed the mix of licenses.

> 2. You might consider adding a `CITATION.cff` file (either to the Zenodo registration of the code, or the JOSS paper when finalised).

Thanks for pointing to this feature. The Github integration makes this format interessting. I will will look into it later.

> 3. How does this benchmark against the alternatives out there? I see it is parallelised internally—how do you specify the number of jobs it can spawn? From the source code, it seems you can pass in this as an argument on the command line, but this isn't reference anywhere I can find.

If a command supports parallel processing, the number of cores can be specified with the `-n` or `--njobs` flag, see for example the CLI help of `yam correlate -h`. Commands correlate, stack and stretch support multiprocessing. By default all available cores are used. I did not run any benchmarks against other tools. I added text about `--njobs`  in the *How to use yam* section of the documentation.

> **Manuscript** I think it would be useful, for the layperson, to highlight what exactly is the value of monitoring relative velocity changes. I also don’t understand in what sense this competes with MSNoise? What do you mean by off-line?

I added some highlights to the statement of need. I see now that off-line could be misinterpreted. I reworded that part. yam is an alternative to MSNoise. I have tried to list some of the differences. Other than that, it is a completely different code base and codes could be used for verification.

> You might specify a reference for the stretching procedure.

Done.

> I like the human-readable, static config file approach.

I agree, that is one of yam's strong points.

> And are graduate students not researchers? Perhaps you mean it is designed for research and teaching purposes.

Of course you are right. I think it is useful for bachelor and master theses, too. I added that point. It could also be useful for teaching, but I think only in very specialized courses.

> **Code** **Platform specs** **OS:** macOS 12.5.1 **Arch:** arm64 (M1 chip) **Python:** 3.9.12
>
> **Download, installation, testing** I was able to download and install `yam` from both the source repository and PyPI. Running the command-line test utility (`yam-runtests`) failed on 1 test. This seemed to be solely timing related, and not to do with the functionality of the code itself. By commenting out lines 173 and 174 and re-installing, I was able to run the tests with no further issues.

That is interesting, I expect that the script loading time on your platform is relatively high. Therefore the relative time saving when processing the small finalized test dataset is not large enough to pass the test. I relaxed the test a bit.

> **Tutorials** I’ll preface this by saying I am not good at following instructions in a linear fashion and will often glaze over some things, run commands, then go back and try figure out why things happened the way they did….
>
> 1. Clarify at the top of the tutorial that the commands are being run in a terminal

Done.

> 2. Perhaps be explicit and include two additional lines at the top of the box:
>        a.	`mkdir yam_tutorial; cd yam_tutorial`
>        b.	`yam create --tutorial`
>        I’ll admit I glazed over the ‘into an empty directory’ and just leapt straight in.

Done.

> 3. `yam plot stations` does not work on Mac M1 when building from source. This is due to `cartopy` not being listed in the requirements. `cartopy` cannot be installed from PyPI (on my M1 system at least), but it does work with `conda`. There are no issues if you the install instructions as specified It might be worth adding a note explicitly stating `cartopy` must be installed via conda, even if building from source?

I added a comment to the corresponding command in the tutorial. Yes, cartopy is not installed automatically, because it is no hard dependency. I do not want to go into too much detail about how to install cartopy on a particular platform. Actually, it is only used indirectly via ObsPy.

> 4. Was very easy to grok how to set things up to run on a locally hosted data set, specifying an archive structure etc. I really like that you can create a collection of configuration recipes without needing to maintain multiple config JSONs. The keys can become a little unwieldy, but after a couple of runs editing things I’m pretty happy with them.

:+1:

> 5. I understood what was happening a lot better when I ran through each section in conjunction with the `conf.json` file. I would recommend people have it open while running through the example sequence of commands.

I added your recommendation to the text prefacing the tuturial.

> 6. I echo @hfmark’s comment—" the human-readable config files and obspy-readable output files [a]re a key difference” that should be highlighted strongly in the manuscript.

I agree. Do you think this is sufficiently highlighted in the current version of the manuscript?

> The test source code contained references to a different package, by the way:
>
> ```
> yam/tests/__init__.py: Tests for the rf package.
> ```

I fixed it.

> I'm looking forward to finding time to apply `yam` to a more extensive swathe of the seismic data I've worked with!
>
> I'll finalise the checklist, but I'm in principle ok to say this is ready to go, once you've had a chance to get back to me on some of the above points. Thanks for your patience!

I will be happy if the package turns out to be useful for you after you have had to dig into it in your review.
