### 2023-05-01 #5267 @elbeejay
@trichter those revisions and responses sound good to me. @hemmelig please have a look when you get a chance and check off your final reviewer checklist items for completeness.

@trichter once that's settled I'll be asking you to do a few small steps to finalize the metadata that will be associated with the JOSS article, and then we'll recommend it for publication.
