### 2023-05-03 #5267 @elbeejay
Great thanks @hemmelig.

@trichter I'll follow up later today with the next steps. But roughly they will involve archiving the version of record and finalizing the metadata (version, archive DOI, etc) that will be associated with the manuscript.
