### 2023-05-03 #5267 @elbeejay
Hi @trichter, I have some minor suggestions for the text itself:

Line 14: "The technique may allow to draw conclusions..." - This should be revised, potentially to "The technique may allow _one_ to draw conclusions..."

Line 26: I suggest swapping "completed datasets" for "complete datasets"

Line 35: "unlike to a web interface..." should be revised, one option is "unlike _the_ web interface..."

Line 41: I don't have a real suggestion here, but I just wanted to echo @hemmelig and note that even after the revision it seems like there is some distinction being drawn between researchers and bachelor and masters theses. Up to you if you'd like to change this, just wanted to provide another opinion.

Once you are happy with the final manuscript @trichter, I need you to do the following:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) - for you this is just updating the archive to the latest versions
- [ ] (please consider) Re-releasing the package on PyPi so that there is a pip-installable version of the package corresponding to what is going to be the JOSS-publication version of record
