### 2023-05-03 #5267 @trichter
I went through the checklist. I created a new release `v0.7.2` on pypi, github and zenodo. The zenodo doi is https://doi.org/10.5281/zenodo.3953646 (pointing to the latest version, preferably) or https://doi.org/10.5281/zenodo.7893762 for version v0.7.2.
