### 2023-05-04 #5267 @elbeejay
Great, thanks @trichter, I will be recommending your manuscript for publication in JOSS now. The next step will be for an EiC to look things over and actually do the publishing - this should happen within the next week.

Many thanks again to @hfmark, @hemmelig, and @ThomasMGeo for their efforts reviewing this submission.
