### 2023-05-09 #5267 @kthyng
Congrats to @trichter on your new publication! Many thanks to editor @elbeejay and reviewers @hfmark, @hemmelig, and @ThomasMGeo for your time, hard work, and expertise!!
